redo vec_index since flann doesnt copy data
