+++ src/lib.rs
@@ -20,6 +20,7 @@ mod parameters;
 #[cfg(test)]
 mod tests;
 mod vec_index;
+mod slice_index;
 
 pub use enums::{Algorithm, CentersInit, Checks, DistanceType, LogLevel};
 pub use generic_array::typenum;
@@ -27,3 +28,4 @@ pub use index::Index;
 pub use indexable::Indexable;
 pub use parameters::Parameters;
 pub use vec_index::VecIndex;
+pub use slice_index::{SliceIndex, FlannError};
