+++ src/parameters.rs
@@ -1,5 +1,6 @@
 use enums::{Algorithm, CentersInit, Checks, DistanceType, LogLevel};
 use raw;
+use std::os::raw::c_long;
 
 #[derive(Debug, Clone)]
 pub struct Parameters {
@@ -23,7 +24,7 @@ pub struct Parameters {
     pub key_size: u32,
     pub multi_probe_level: u32,
     pub log_level: LogLevel,
-    pub random_seed: i32,
+    pub random_seed: c_long,
     pub distance_type: DistanceType,
     pub distance_order: i32,
 }
